matlab_mex: use generic sim_set_ext_fun for all models by using matlab eval & c preprocessing for templating
