+++ interfaces/acados_matlab/sim_compile_mex_model_dep.m
@@ -1,4 +1,4 @@
-function sim_compile_mex_model_dep(model_struct, opts_struct)
+function C_sim_ext_fun = sim_compile_mex_model_dep(C_sim, C_sim_ext_fun, model_struct, opts_struct)
 
 % get acados folder
 acados_folder = getenv('ACADOS_INSTALL_DIR');
@@ -13,58 +13,157 @@ acados_matlab_lib_path = ['-L' acados_folder, '/interfaces/acados_matlab/'];
 model_lib_path = ['-L', pwd, '/build'];
 
 %% select files to compile
-mex_files = {};
+mex_files = {[acados_mex_folder, 'sim_set_ext_fun_dyn.c']};
+set_fields = {};
+mex_fields = {};
+fun_names = {};
+mex_names = {};
 % dynamics
 if (strcmp(opts_struct.method, 'erk'))
-	mex_files = {mex_files{:}, ...
-		[acados_mex_folder, 'sim_set_ext_fun_dyn_expl.c'] ...
+
+	set_fields = {set_fields{:} ...
+		'expl_ode_fun' ...
+		'expl_vde_for' ...
+		'expl_vde_adj' ...
+		'expl_ode_hes' ...
 		};
+	mex_fields = {mex_fields{:} ...
+		'dyn_expl_ode_fun' ...
+		'dyn_expl_vde_for' ...
+		'dyn_expl_vde_adj' ...
+		'dyn_expl_ode_hes' ...
+		};
+	fun_names = {fun_names{:} ...
+		'sim_model_dyn_expl_ode_fun' ...
+		'sim_model_dyn_expl_vde_for' ...
+		'sim_model_dyn_expl_vde_adj' ...
+		'sim_model_dyn_expl_ode_hes' ...
+		};
+	mex_names = {mex_names{:} ...
+		'sim_set_ext_fun_dyn_expl_ode_fun' ...
+		'sim_set_ext_fun_dyn_expl_vde_for' ...
+		'sim_set_ext_fun_dyn_expl_vde_adj' ...
+		'sim_set_ext_fun_dyn_expl_ode_hes' ...
+		};
+
 elseif (strcmp(opts_struct.method, 'irk'))
-	mex_files = {mex_files{:}, ...
-		[acados_mex_folder, 'sim_set_ext_fun_dyn_impl.c'] ...
+
+	set_fields = {set_fields{:} ...
+		'impl_ode_fun' ...
+		'impl_ode_fun_jac_x_xdot' ...
+		'impl_ode_jac_x_xdot_u' ...
+		'impl_ode_hess' ...
 		};
+	mex_fields = {mex_fields{:} ...
+		'dyn_impl_ode_fun' ...
+		'dyn_impl_ode_fun_jac_x_xdot' ...
+		'dyn_impl_ode_jac_x_xdot_u' ...
+		'dyn_impl_ode_hess' ...
+		};
+	fun_names = {fun_names{:} ...
+		'sim_model_dyn_impl_ode_fun' ...
+		'sim_model_dyn_impl_ode_fun_jac_x_xdot' ...
+		'sim_model_dyn_impl_ode_jac_x_xdot_u' ...
+		'sim_model_dyn_impl_ode_hess' ...
+		};
+	mex_names = {mex_names{:} ...
+		'sim_set_ext_fun_dyn_impl_ode_fun' ...
+		'sim_set_ext_fun_dyn_impl_ode_fun_jac_x_xdot' ...
+		'sim_set_ext_fun_dyn_impl_ode_jac_x_xdot_u' ...
+		'sim_set_ext_fun_dyn_impl_ode_hess' ...
+		};
+
 elseif (strcmp(opts_struct.method, 'irk_gnsf'))
-	mex_files = {mex_files{:}, ...
-		[acados_mex_folder, 'sim_set_ext_fun_dyn_gnsf.c'] ...
+
+	set_fields = {set_fields{:} ...
+		'gnsf_f_lo_fun_jac_x1k1uz' ...
+		'gnsf_get_matrices_fun' ...
+		'gnsf_phi_fun' ...
+		'gnsf_phi_fun_jac_y' ...
+		'gnsf_phi_jac_y_uhat' ...
 		};
+	mex_fields = {mex_fields{:} ...
+		'dyn_gnsf_f_lo_fun_jac_x1k1uz' ...
+		'dyn_gnsf_get_matrices_fun' ...
+		'dyn_gnsf_phi_fun' ...
+		'dyn_gnsf_phi_fun_jac_y' ...
+		'dyn_gnsf_phi_jac_y_uhat' ...
+		};
+	fun_names = {fun_names{:} ...
+		'sim_model_dyn_gnsf_f_lo_fun_jac_x1k1uz' ...
+		'sim_model_dyn_gnsf_get_matrices_fun' ...
+		'sim_model_dyn_gnsf_phi_fun' ...
+		'sim_model_dyn_gnsf_phi_fun_jac_y' ...
+		'sim_model_dyn_gnsf_phi_jac_y_uhat' ...
+		};
+	mex_names = {mex_names{:} ...
+		'sim_set_ext_fun_dyn_gnsf_f_lo_fun_jac_x1k1uz' ...
+		'sim_set_ext_fun_dyn_gnsf_get_matrices_fun' ...
+		'sim_set_ext_fun_dyn_gnsf_phi_fun' ...
+		'sim_set_ext_fun_dyn_gnsf_phi_fun_jac_y' ...
+		'sim_set_ext_fun_dyn_gnsf_phi_jac_y_uhat' ...
+		};
+
 else
 	fprintf('\ncodegen_model: method not supported: %s\n', opts_struct.method);
 end
 
-if is_octave()
-	if exist('build/cflags_octave.txt')==0
-		diary 'build/cflags_octave.txt'
-		diary on
-		mkoctfile -p CFLAGS
-		diary off
-		input_file = fopen('build/cflags_octave.txt', 'r');
-		cflags_tmp = fscanf(input_file, '%[^\n]s');
-		fclose(input_file);
-		cflags_tmp = [cflags_tmp, ' -std=c99 -fopenmp'];
-		input_file = fopen('build/cflags_octave.txt', 'w');
-		fprintf(input_file, '%s', cflags_tmp);
-		fclose(input_file);
-	end
-	input_file = fopen('build/cflags_octave.txt', 'r');
-	cflags_tmp = fscanf(input_file, '%[^\n]s');
-	fclose(input_file);
-	setenv('CFLAGS', cflags_tmp);
-end
+% compile mex files
+if (strcmp(opts_struct.compile_mex, 'true'))
 
-%% get pointers for external functions in model
-for ii=1:length(mex_files)
-	disp(['compiling ', mex_files{ii}])
 	if is_octave()
-%		mkoctfile -p CFLAGS
-		mex(acados_include, acados_interfaces_include, acados_lib_path, acados_matlab_lib_path, model_lib_path, '-lacados', '-lhpipm', '-lblasfeo', '-lsim_model', mex_files{ii});
-	else
-		mex(mex_flags, 'CFLAGS=\$CFLAGS -std=c99 -fopenmp', acados_include, acados_interfaces_include, acados_lib_path, acados_matlab_lib_path, model_lib_path, '-lacados', '-lhpipm', '-lblasfeo', '-lsim_model', mex_files{ii});
+		if exist('build/cflags_octave.txt')==0
+			diary 'build/cflags_octave.txt'
+			diary on
+			mkoctfile -p CFLAGS
+			diary off
+			input_file = fopen('build/cflags_octave.txt', 'r');
+			cflags_tmp = fscanf(input_file, '%[^\n]s');
+			fclose(input_file);
+			cflags_tmp = [cflags_tmp, ' -std=c99 -fopenmp'];
+			input_file = fopen('build/cflags_octave.txt', 'w');
+			fprintf(input_file, '%s', cflags_tmp);
+			fclose(input_file);
+		end
+	%	input_file = fopen('build/cflags_octave.txt', 'r');
+	%	cflags_tmp = fscanf(input_file, '%[^\n]s');
+	%	fclose(input_file);
+	%	setenv('CFLAGS', cflags_tmp);
 	end
+
+	%% get pointers for external functions in model
+	for ii=1:length(mex_names)
+
+		disp(['compiling ', mex_names{ii}])
+		if is_octave()
+	%		mkoctfile -p CFLAGS
+			input_file = fopen('build/cflags_octave.txt', 'r');
+			cflags_tmp = fscanf(input_file, '%[^\n]s');
+			fclose(input_file);
+			cflags_tmp = [cflags_tmp, ' -DSET_FIELD=', set_fields{ii}];
+			cflags_tmp = [cflags_tmp, ' -DMEX_FIELD=', mex_fields{ii}];
+			cflags_tmp = [cflags_tmp, ' -DFUN_NAME=', fun_names{ii}];
+			setenv('CFLAGS', cflags_tmp);
+			mex(acados_include, acados_interfaces_include, acados_lib_path, acados_matlab_lib_path, model_lib_path, '-lacados', '-lhpipm', '-lblasfeo', '-lsim_model', mex_files{1});
+			system(['mv sim_set_ext_fun_dyn.mex ', mex_names{ii}, '.mex']);
+		else
+			mex(mex_flags, 'CFLAGS=\$CFLAGS -std=c99 -fopenmp', ['-DSET_FIELD=', set_fields{ii}], ['-DMEX_FIELD=', mex_fields{ii}], ['-DFUN_NAME=', fun_names{ii}], acados_include, acados_interfaces_include, acados_lib_path, acados_matlab_lib_path, model_lib_path, '-lacados', '-lhpipm', '-lblasfeo', '-lsim_model', mex_files{1});
+			system(['mv sim_set_ext_fun_dyn.mexa64 ', mex_names{ii}, '.mexa64']);
+		end
+
+	end
+
+	if is_octave()
+		system(['mv -f *.o build/']);
+		system(['mv -f *.mex build/']);
+	else
+		system(['mv -f *.mexa64 build/']);
+	end
+
 end
 
-if is_octave()
-	system(['mv -f *.o build/']);
-	system(['mv -f *.mex build/']);
-else
-	system(['mv -f *.mexa64 build/']);
+% call mex files
+for ii=1:length(mex_names)
+	C_sim_ext_fun = eval([mex_names{ii}, '(C_sim, C_sim_ext_fun, model_struct, opts_struct)']);
 end
+
